requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
Pillow==10.1.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.25
requests==2.31.0
whitenoise==6.6.0
django-filter==23.5
gunicorn==21.2.0